requirements.txt
google-cloud-bigquery==3.3.1
google-cloud-language==2.5.2
google-cloud-translate==3.8.1